requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据处理
pandas==2.1.3
openpyxl==3.1.2
numpy==1.26.2

# PDF处理
pdfplumber==0.10.3
reportlab==4.0.7

# NLP & 向量模型
FlagEmbedding==1.2.3
jieba==0.42.1
scikit-learn==1.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# 向量数据库
chromadb==0.4.18

# 数据库
sqlalchemy==2.0.23

# 前端
streamlit==1.28.2
plotly==5.18.0

# 工具
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.2
pyyaml==6.0.1

# PyTorch (CPU版本，轻量级)
torch==2.0.0
torchvision==0.15.0
--extra-index-url https://download.pytorch.org/whl/cpu

# LLM (OpenAI 或兼容)
openai>=1.40.0
